Refactor fixtures in facade tests
