+++ server/nlp_ssa/models/article_data/tests/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from models.article_data import ArticleDataFacade
+
+
+@pytest.fixture
+def article_data_facade(mock_db_session):
+    return ArticleDataFacade(db_session=mock_db_session)
