+++ server/nlp_ssa/models/article_data/tests/test_facade.py
@@ -10,18 +10,7 @@ from models.article_data import (
 )
 from models.article_data.factories import ArticleDataFactory
 from models.stock.factories import StockFactory
-
-
-@pytest.fixture(autouse=True)
-def mock_facade_now(mocker, mock_utcnow):
-    return mocker.patch(
-        "nlp_ssa.models.article_data.facade.arrow.utcnow", return_value=mock_utcnow
-    )
-
-
-@pytest.fixture
-def article_data_facade(mock_db_session):
-    return ArticleDataFacade(db_session=mock_db_session)
+from utils.testing_mocks import get_mock_utcnow
 
 
 def test_get_one_by_id(article_data_facade, mock_db_session):
@@ -97,9 +86,7 @@ def test_get_all_by_stock_symbol_no_results(article_data_facade):
     assert results == []
 
 
-def test_create_or_update_new_article_data(
-    article_data_facade, mock_db_session, mock_utcnow, mock_facade_now
-):
+def test_create_or_update_new_article_data(article_data_facade, mock_db_session):
     StockFactory(quote_stock_symbol="NTDOF")
     mock_db_session.commit()
 
@@ -125,9 +112,7 @@ def test_create_or_update_new_article_data(
     assert result.source_url == article_data_dict["source_url"]
     assert result.raw_content == article_data_dict["raw_content"]
     assert result.sentence_tokens == article_data_dict["sentence_tokens"]
-    # TODO: not sure why the mocks in the test fixtures aren't working :']
-    # assert result.created_at == mock_utcnow
-    # assert result.updated_at == mock_utcnow
+    # TODO: find better way to mock server 'now' function
     assert isinstance(result.created_at, arrow.Arrow)
     assert isinstance(result.updated_at, arrow.Arrow)
 
@@ -176,6 +161,9 @@ def test_create_or_update_existing_article_data(article_data_facade, mock_db_ses
     assert result.source_url == mock_article_data.source_url
     assert result.raw_content == updated_article_data_dict["raw_content"]
     assert result.sentence_tokens == updated_article_data_dict["sentence_tokens"]
+    assert result.created_at == get_mock_utcnow()
+    # TODO: find better way to mock server 'now' function
+    assert isinstance(result.updated_at, arrow.Arrow)
 
     assert result.author == ""
     assert result.last_updated_date is None
