+++ server/nlp_ssa/models/sentiment_analysis/tests/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from models.sentiment_analysis import SentimentAnalysisFacade
+
+
+@pytest.fixture
+def sentiment_analysis_facade(mock_db_session):
+    return SentimentAnalysisFacade(db_session=mock_db_session)
