+++ server/nlp_ssa/models/stock/factories.py
@@ -1,4 +1,3 @@
-import arrow
 import factory
 import uuid
