+++ server/nlp_ssa/models/stock/tests/conftest.py
@@ -1,0 +1,8 @@
+import pytest
+
+from models.stock import StockFacade
+
+
+@pytest.fixture
+def stock_facade(mock_db_session):
+    return StockFacade(db_session=mock_db_session)
